MueLu: add new RAPFactory unit test
This commit adds a unit test for the option
"rap: fix zero diagonals".  This option inserts
ones on the main diagonal in rows that are either
empty or have a zero on the diagonal.

With a Kokkos OpenMP node, this option was causing
such rows to have multiple diagonal entries, one
zero and one nonzero.

I've verified that the unit test catches the problem.
It should now pass because MueLu_RAPFactory has been
patched.

A big thank you to Christian G., who first identified
the multiple diagonal entry issue and proposed a fix.