MueLu: fix option to replace zero diag elements
The implementation of the option "rap: fix zero diagonals"
was creating multiple entries in a row with the same column index.
This was causing problems downstream. For example, direct solvers
would detect a zero pivot.

Not calling "resumeFill" avoids this problem (for unknown reasons).
This commit also switches to the three-matrix variant of matrix addition.

Note that the the underlying issue is not fixed.  One clue -- this only
manifests with the Kokkos OpenMP node type, not the serial node type.

A big thank you to Christian G., who first identified
the multiple diagonal entry issue and proposed a fix.