Accidentally set OVERALL_NUM_MPI_PROCS to 10 instead of 1 (#3133)
This did not impact how the test was run, it just told CTest that these tests
were using 10 MPI processes so it would not have run other tests at the same
time.  So this just wasted a little bit of wall-clock time running the tests
(but not much since these tests run pretty fast).