Fix calling of get-changed-trilinos-packages.sh and other improvements (#3133, #3265)
Now calling this script will select the correct set of file changes and
correct set of enables as part the goals of #3133.  I have tested this manally
on several different PRs and verified that the correct set of enables are
being selected.

I did a few other things while editing the script:

* The script can now be run on a different scratch Trilinos repo.  This allows
  it to be manually tested without messing up the status of your main Trilinos
  git repo (see details in PR #3258)

* The Trilinos git repo where the PR branch is merged is now automatically
  updated to the correct state, no matter what its prior state is before it is
  run.  The only requirement is that the 'origin' remote is set correctly.
  None of the other local branches have to be exist or set correctly at all.
  See the new comments at the top.

* Make all formating if statements the same with the same indentation.

* Removed code that was not used or did nothing.

Without this change, the auto PR tester will test more package that is it
supposed to.