Teuchos: ensure all YAML params get root name
This is a fix of a previous change, which would ensure
that if the YAML file read:

mycode:
  param1: foo

That the top-level ParameterList would be named "mycode"
instead of "ANONYMOUS".
This change ensures that that is propagated to the names
of all subparameters, e.g. "mycode -> param1" instead of
"ANONYMOUS -> param1".

This prevents user confusion in seeing "ANONYMOUS"
despite never writing it in their input file.