Tempus:  Add potential full-space optimization interface.
This commit adds a transient full-space optimization interface to be
used with packages such as ROL with its DynamicConstraint interface.  An
implementation of the interface is provided by BackwardEuler only
(currently).  A test that the interface computes what one would expect
for BackwardEuler is also provided.  The idea is to then implement ROL's
DynamicConstraint interface using Tempus, and then continue to develop
and refine the interface.