MueLu: Xpetra: add threshold for zero diag fix
This commit changes the behavior of the option "rap: fix zero diagonals".
It now flags diagonal entries that are less than or equal to
ScalarTraits<SC>::eps(), rather than a hard zero.  There is also a new option
"rap: fix zero diagonals threshold" that allows one to change the threshold.