Framework: Enable extended options in parameterized tests (#3555)
@trilinos/framework 	

@jwillenbring 	
@prwolfe 	

This is an update to the script for the Tpetra request to enable COMPLEX_DOUBLE in at least one of the nightly tests, with movement up to Clean testing once the kinks are sorted out.  That requires the modifications to go into the parameterized tests.

This modification enables us to set two additional parameters in Jenkins:
- `JENKINS_Trilinos_EXTRA_CONFIGURE_OPTIONS` - Takes a string in the form "-DTrilinos_ENABLE_..." "-DTrilinos_ENABLE_..." and inserts this into the `EXTRA_CONFIGURE_OPTIONS` in the testing.
- `JENKINS_BUILD_TAG` - A string that will get appended to the testing label in CDash.  
  - If this is empty, then nothing gets added.  
  - If it's not empty, then "_${JENKINS_BUILD_TAG}" will get appended to the test label in CDash.
  - Note: Leading underscores in the value of this will get removed so there is always only 1 leading underscore.

This update fixes a bug with the previous attempt which didn't take into account the way we launch these jobs in Jenkins.  

@jwillenbring
@prwolfe 
You can review it and approve if you get a chance, but I have a test running currently and want it to finish before merging to make sure it does the right thing.
- [CDash output from test][1]
- [Jenkins Job][2]

[1]: https://testing.sandia.gov/cdash/index.php?project=Trilinos&parentid=4000539
[2]: https://ascic-jenkins.sandia.gov/job/trilinos-folder/job/Trilinos_generic_nightly_test_wcmclen/20/