Disable SuperLU_Node and ROL in non-SPARC builds for now (TRIL-212)
For now, we will disable these packages not being used by EMPIRE for now so
that we don't introduce any new failing tests in the existing ATDM Trilinos
builds.  But as we get the SPARC ATDM Trilinos configuration building on more
machines and we test SPARC and EMPIRE against the fuller SPARC configuration,
then ShyLU_Node and ROL will get enabled one machine at a time in an orderly
fashion.