MueLu: fix test hangs under CUDA
MueLu's unit tests for Create*PetraPreconditioner were incorrectly
providing a scalar nullspace for a matrix with numPDEs=3.  This was
causing a hang in the test down in KokkosKernel's SpGEMM on CUDA.

This commit fixes the unit tests in question by specifying a nullspace
with as many columns as numPDEs, adds a check in HierarchyManager that
the nullspace dimension is at least as large as numPDEs, and adds a unit
test to exercise this new check.

Thanks to CMS for his help in diagnosing this.

Fixes github issues #3655 and #2674.