Stokhos:  Replace STEQR with PTEQR for issue #3542.
The LAPACK STEQR function, which computes eigenvalues/vectors of
symmetric, tridiagonal matrices, seg faults on the IBM Power systems for
some unknown reason.  However PTEQR, which does the same for SPD
systems, appears to work.  So I replaced to the call to STEQR in the
recurrence basis implementation with PTEQR.  This was more complicated
than it sounds because the recurrence matrix is not positive definite.
So I had to put in a shift to make it so for PTEQR.  This required
loosening a few tight numerical tolerances in some tests.