Sacado:  Fix build errors reported in issue #3845.
With Sacado_ENABLE_HIERARCHICAL_DFAD=ON, as was turned on in PR #3800 in
ATDM builds, a build error was seen in the unit tests that disable the
View specializations for Fad, since is_dynrankview_fad_contiguous<> is
used without being defined.  I moved the default definition of this
trait class so that it is always defined, and fix the build errors in my
CUDA build.