Disable test SEACASIoss_Utst_structured_decomp_MPI_1 and exec build for cee-rhel6 intel builds (#3891)
There is currently a strange link error.  Only this one test execuable is
impacted.

This test links and runs just fine in the other 'cee-rhel6' builds so
disabling it for now in these Intel builds is not so terrible and this can
still be fixed offline.  By disabling it for now, we remove a lot of red spam
from the CDash site and the ATDM Trilinos summary emails (see #2933).