Ifpack2: fixed AdditiveSchwarz test failure
Fixed 3 unit tests under AdditiveSchwarz that failed with
deprecated code off. In RILUK and Amesos2Wrapper,
Tpetra::CrsMatrix::doImport() was being used to copy
a ReorderFilter to a CrsMatrix. The import's src/dst maps were the same.
But, that now requires the source matrix to support getGraph,
which ReorderFilter does not.